unsigned load opcode; refactor bsg_cache regression using python trace assembler
